web/requirements.txt
flask
flask-cors
paho-mqtt
pymongo
simplejson
watchdog